modules/libjar/test/mochitest/test_bug1173171.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1173171
-->
<head>
  <title>Test for Bug 1173171</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>

<iframe id="testFrame"></iframe>

<pre id="test">
<script class="testbody" type="text/javascript;version=1.6">

/** Test for Bug 1173171 **/

SimpleTest.waitForExplicitFinish();

// __setPref(key, value)__.
// Set a pref value asynchronously, returning a prmoise that resolves
// when it succeeds.
let pushPref = function (key, value) {
  return new Promise(function(resolve, reject) {
    SpecialPowers.pushPrefEnv({"set": [[key, value]]}, resolve);
  });
};

// __xhr(method, url, responseType__.
// A simple async XMLHttpRequest call.
// Returns a promise with the response.
let xhr = function (method, url, responseType) {
  return new Promise(function (resolve, reject) {
    let xhr = new XMLHttpRequest();
    xhr.open(method, url, true);
    xhr.onload = function () {
      resolve(xhr.response);
    };
    xhr.responseType = responseType;
    xhr.send();
  });
};

let localAddress = "file://bug403331.zip!/test.html",
    remoteAddress = "http://mochi.test:8888/tests/modules/libjar/test/mochitest/bug403331.zip!/test.html";

// Test behavior when blocking is deactivated and activated, and when a jar file is local or remote.
spawnTask(function* () {
  for (let blocking of [false, true]) {
    yield pushPref("network.jar.block-remote-files", blocking);
    for (let remote = [false, true]) {
      let address = remote ? remoteAddress : localAddress;
      try {
        let response = yield xhr("GET", address, "document");
        ok(!blocking || !remote, "Local files should not be blocked. Neither should remote files when blocking is off.");
      } catch (e) {
        ok(blocking && remote, "Remote files should be blocked when 'network.jar.block-remote-files' pref is true.");
      }
    }
  }
  SimpleTest.finish();
});

</script>
</pre>

</body>
</html>
